refactor: co-locate component tests next to their source files
Move image-carousel.test.tsx and image-lightbox.test.tsx out of
__tests__/ to sit alongside their components, matching the existing
project convention. Remove the now-empty __tests__ directory.
