+++ src/components/image-carousel.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi } from "vitest"
 import { render, screen } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
 import type { ImageRead } from "@/api/generated/types"
-import { ImageCarousel } from "../image-carousel"
+import { ImageCarousel } from "./image-carousel"
 
 function makeImage(overrides: Partial<ImageRead> = {}): ImageRead {
   return {
